+++ src/main/java/com/example/monitoring/core/api/auth/AuthenticationRequest.java
@@ -1,4 +1,4 @@
-package com.example.monitoring.core.model.auth;
+package com.example.monitoring.core.api.auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
