+++ src/main/java/com/example/monitoring/core/api/auth/RegisterRequest.java
@@ -1,4 +1,4 @@
-package com.example.monitoring.core.model.auth;
+package com.example.monitoring.core.api.auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,6 +9,8 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class AuthenticationResponse {
-    private String token;
+public class RegisterRequest {
+    private String deviceId;
+    private String deviceType;
+    private String password;
 }
