+++ src/main/java/com/example/monitoring/core/api/config/SecurityConfiguration.java
@@ -1,4 +1,4 @@
-package com.example.monitoring.core.model.config;
+package com.example.monitoring.core.api.config;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
