+++ src/main/java/com/example/monitoring/core/api/payload/PayloadRepository.java
@@ -1,4 +1,4 @@
-package com.example.monitoring.core.model.payload;
+package com.example.monitoring.core.api.payload;
 
 import org.springframework.data.jpa.repository.JpaRepository;
